Initial work on switching to Typescript
